# Speed up JSON mappings