--- a/frontend/javascripts/oxalis/model/reducers/dataset_reducer.ts
+++ b/frontend/javascripts/oxalis/model/reducers/dataset_reducer.ts
@@ -43,7 +43,6 @@ function DatasetReducer(state: OxalisState, action: Action): OxalisState {
           activeMappingByLayer: createDictWithKeysAndValue(segmentationLayerNames, () => ({
             mappingName: null,
             mapping: null,
-            mappingKeys: null,
             mappingColors: null,
             hideUnmappedIds: false,
             mappingStatus: MappingStatusEnum.DISABLED,