Also, no longer maintain the extra mappingKeys array which was only needed, because Object.keys was very slow for large objects. I benchmarked the old code, the new code, and the new code maintaining the extra mappingKeys array (all with deepFreeze disabled!) using a large JSON mapping with 1.5M entries:

Object mapping with extra keys array (old):
    Mapping creation: ~4600 ms
    Mapping texture creation: ~60 ms

Map mapping with extra keys array (intermediate):
    Mapping creation: ~1160 ms
    Mapping texture creation: ~460 ms

Map mapping (new):
    Mapping creation: ~590 ms
    Mapping texture creation: ~1070 ms
    
Overall a nice 3x speedup for larger JSON mappings. However, I mainly did this in preparation for https://github.com/scalableminds/webknossos/pull/7654 where this frontend mapping functionality will be used extensively.

### URL of deployed dev instance (used for testing):
- https://___.webknossos.xyz

### Steps to test:
- Open a dataset with JSON mappings (e.g. ROI2017_wkw) and check that they still work (and are faster than before)
- Test that merger mode still works as expected

### Issues:
- contributes to https://github.com/scalableminds/webknossos/pull/7654

------
(Please delete unneeded items, merge only when none are left open)
- [x] Updated [changelog](../blob/master/CHANGELOG.unreleased.md#unreleased)
- [x] Removed dev-only changes like prints and application.conf edits
- [x] Considered [common edge cases](../blob/master/.github/common_edge_cases.md)